+++ src/cylestio_monitor/__init__.py
@@ -55,7 +55,7 @@ from cylestio_monitor.utils.trace_context import TraceContext
 # Import the API client module to make it available
 from . import api_client
 
-__version__ = "0.1.10"
+__version__ = "0.1.11"
 
 __all__ = [
     "start_monitoring",